disclaude/requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
lark-oapi>=1.2.0
aiohttp>=3.9.0
